Fix and/or silence a bunch of MSVC warnings.
[libmspack-xenia] lzxd.c: Cast pointer differences to int.

[microtar] microtar.c: fwrite() and fread() return size_t.

[libcachecommon] CacheKeys.hpp: Silence C4834 and gcc -Wunused-result
for cacheKeys.c_str().

[libromdata] ELF.cpp: Cast val_dtag[] to uint32_t.

[libromdata] EXE_NE.cpp: Cast ao::uvector<> sizes to uint32_t.

[libromdata/tests] microtar_zstd.c: Cast toCopy to unsigned when passing
it to microtar functions.

[librpfile] RpFile_scsi_win32.cpp: Windows 8.1 SDK's ntddscsi.h doesn't
properly declare _NV_SEP_WRITE_CACHE_TYPE as an enum.

[librptext/tests] TextFuncsTest: Silence warnings about implicit
conversoin from double to unsigned int by only doing integer
calculations. (It's done at compile time, not runtime.)

[librptexture] PowerVR3, ValveVTF3: Explicitly cast some size_t
calculations to unsigned int.

[win32] KeyManagerTab: Explicitly cast the return value from
std::count_if() to int.

+++ extlib/libmspack-xenia/lzxd.c
@@ -409,7 +409,7 @@ int lzxd_decompress(struct lzxd_stream *lzx, off_t out_bytes) {
   if (lzx->error) return lzx->error;
 
   /* flush out any stored-up bytes before we begin */
-  i = lzx->o_end - lzx->o_ptr;
+  i = (int)(lzx->o_end - lzx->o_ptr);
   if ((off_t) i > out_bytes) i = (int) out_bytes;
   if (i) {
     if (lzx->sys->write(lzx->output, lzx->o_ptr, i) != i) {
@@ -779,7 +779,7 @@ int lzxd_decompress(struct lzxd_stream *lzx, off_t out_bytes) {
         rundest = &window[window_posn];
         window_posn += this_run;
         while (this_run > 0) {
-          if ((i = i_end - i_ptr) == 0) {
+          if ((i = (int)(i_end - i_ptr)) == 0) {
             READ_IF_NEEDED;
           }
           else {
